Refactor: Move email templates to components directory
